added wrapper for storage to use in auth service
