+++ internal/services/auth/service.go
@@ -1,38 +1,22 @@
 package auth
 
 import (
-	"context"
 	"log/slog"
 	"time"
 
-	"github.com/Onnywrite/grpc-auth/internal/models"
+	"github.com/Onnywrite/grpc-auth/internal/services/auth/wrap"
 )
-
-type Storage interface {
-	SaveUser(ctx context.Context, user *models.User) (*models.SavedUser, error)
-	User(ctx context.Context, user models.UserIdentifier) (*models.SavedUser, error)
-
-	SaveSignup(ctx context.Context, signup models.Signup) (*models.SavedSignup, error)
-	Signup(ctx context.Context, userId, serviceId int64) (*models.SavedSignup, error)
-
-	SaveSession(ctx context.Context, session *models.Session) (*models.SavedSession, error)
-	SessionById(ctx context.Context, uuid string) (*models.SavedSession, error)
-	Session(ctx context.Context, session *models.Session) (*models.SavedSession, error)
-	TerminateSession(ctx context.Context, uuid string) error
-	ReviveSession(ctx context.Context, uuid string) error
-	DeleteSession(ctx context.Context, uuid string) error
-}
 
 type AuthService struct {
 	log                                   *slog.Logger
-	db                                    Storage
+	db                                    *wrap.Wrapper
 	tokenTTL, refreshTokenTTL, idTokenTTL time.Duration
 }
 
-func New(logger *slog.Logger, db Storage, tokenTTL, refreshTokenTTL, idTokenTTL time.Duration) *AuthService {
+func New(logger *slog.Logger, db wrap.Storage, tokenTTL, refreshTokenTTL, idTokenTTL time.Duration) *AuthService {
 	return &AuthService{
 		log:      logger,
-		db:       db,
+		db:       wrap.New(logger, db),
 		tokenTTL: tokenTTL,
 	}
 }
